Draft: split makeSketch from Draft.py
