+++ src/Mod/Draft/CMakeLists.txt
@@ -74,6 +74,7 @@ SET(Draft_make_functions
     draftmake/make_polygon.py
     draftmake/make_point.py
     draftmake/make_rectangle.py
+    draftmake/make_sketch.py
     draftmake/make_wire.py
     draftmake/make_wpproxy.py
 )
